# swagger.yaml
basePath: /api/v1
definitions:
...
  handler.MPBindPhoneData:
    properties:
      phone_masked:
        example: 138****8888
        type: string
    type: object
  handler.MPBindPhoneReq:
    properties:
      phone_code:
        example: wx_phone_code_from_getPhoneNumber
        type: string
    type: object
...
  handler.MPCreateOrderData:
    properties:
      amount:
        example: 9900
        type: integer
      order_id:
        example: ORD00000000000000000001
        type: string
      payment_params:
        $ref: '#/definitions/handler.MPPaymentParamsDoc'
      report_id:
        example: 101
        type: integer
    type: object
  handler.MPCreateOrderReq:
    properties:
      amount:
        example: 9900
        type: integer
      report_id:
        example: 101
        type: integer
    type: object
  handler.MPCreateReportData:
    properties:
      created_at:
        example: "2026-04-20T10:00:00Z"
        type: string
      name:
        example: 张三
        type: string
      report_id:
        example: 101
        type: integer
      report_type:
        example: preview
        type: string
      status:
        example: draft
        type: string
    type: object
  handler.MPCreateReportReq:
    properties:
      biology_score:
        example: 75
        type: integer
      chemistry_score:
        example: 80
        type: integer
      chinese_score:
        example: 115
        type: integer
      college_name:
        example: 信息科学技术学院
        type: string
      employment_intention:
        items:
          type: string
        type: array
      english_level:
        example: CET-6
        type: string
      english_score:
        example: 130
        type: integer
      enrollment_year:
        example: 2021
        type: integer
      gaokao_province:
        example: 广东
        type: string
      gaokao_rank:
        example: 5000
        type: integer
      gaokao_score:
        example: 620
        type: integer
      gender:
        example: 男
        type: string
      hukou:
        example: 广东广州
        type: string
      major_name:
        example: 计算机科学与技术
        type: string
      major_satisfaction:
        example: 一般满意
        type: string
      math_score:
        example: 140
        type: integer
      name:
        example: 张三
        type: string
      notes:
        example: 对数据方向感兴趣
        type: string
      physics_score:
        example: 85
        type: integer
      school_name:
        example: 北京大学
        type: string
      study_intention:
        example: 一定要升
        type: string
      study_path_priority:
        items:
          type: string
        type: array
      target_major:
        items:
          type: string
        type: array
      target_work_city:
        items:
          type: string
        type: array
    type: object
...
  handler.MPErrorResp:
    properties:
      code:
        example: 1001
        type: integer
      message:
        example: invalid request
        type: string
      timestamp:
        example: 1713600000
        type: integer
    type: object
...
  handler.MPLoginData:
    properties:
      has_phone:
        example: false
        type: boolean
      is_new_user:
        example: true
        type: boolean
      role:
        enum:
        - user
        - distributor
        example: user
        type: string
      user_info:
        $ref: '#/definitions/handler.MPUserInfoDoc'
    type: object
  handler.MPLoginReq:
    properties:
      distributor_id:
        example: 100
        type: integer
    type: object
...
  handler.MPMeData:
    properties:
      avatar_url:
        example: https://thirdwx.qlogo.cn/...
        type: string
      created_at:
        example: "2026-04-20T10:00:00Z"
        type: string
      distributor_level:
        example: campus
        type: string
      has_phone:
        example: true
        type: boolean
      is_distributor:
        example: false
        type: boolean
      nickname:
        example: 张三
        type: string
      parent_distributor_id:
        example: 100
        type: integer
      phone_masked:
        example: 138****8888
        type: string
      report_count:
        example: 2
        type: integer
      role:
        enum:
        - user
        - distributor
        example: user
        type: string
      unread_message_count:
        example: 3
        type: integer
      user_id:
        example: 42
        type: integer
    type: object
...
  handler.MPOrderDetailData:
    properties:
      amount:
        example: 9900
        type: integer
      created_at:
        example: "2026-04-20T09:55:00Z"
        type: string
      order_id:
        example: ORD00000000000000000001
        type: string
      paid_at:
        example: "2026-04-20T10:00:00Z"
        type: string
      report_id:
        example: 101
        type: integer
      status:
        enum:
        - pending
        - paid
        - failed
        example: paid
        type: string
    type: object
  handler.MPOrderDetailReq:
    properties:
      order_id:
        example: ORD00000000000000000001
        type: string
    type: object
  handler.MPPageDataDoc:
    properties:
      list: {}
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      page_total:
        example: 5
        type: integer
      total:
        example: 100
        type: integer
    type: object
  handler.MPPageReq:
    properties:
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
    type: object
  handler.MPPageResponse:
    properties:
      code:
        example: 0
        type: integer
      data:
        $ref: '#/definitions/handler.MPPageDataDoc'
      message:
        example: ok
        type: string
      timestamp:
        example: 1713600000
        type: integer
      user_info:
        $ref: '#/definitions/handler.MPUserInfoDoc'
    type: object
  handler.MPPaymentParamsDoc:
    properties:
      nonceStr:
        example: abc123xyz
        type: string
      package:
        example: prepay_id=wx20260420...
        type: string
      paySign:
        example: RSA签名字符串...
        type: string
      signType:
        example: RSA
        type: string
      timeStamp:
        example: "1713600000"
        type: string
    type: object
...
  handler.MPReportIDReq:
    properties:
      report_id:
        example: 101
        type: integer
    type: object
  handler.MPReportLinksData:
    properties:
      expires_in:
        example: 3600
        type: integer
      full_h5_url:
        example: "null"
        type: string
      is_paid:
        example: true
        type: boolean
      pdf_url:
        example: "null"
        type: string
      preview_h5_url:
        example: https://cos.example.com/h5/101_preview.html?sign=xxx
        type: string
      report_id:
        example: 101
        type: integer
    type: object
  handler.MPReportListItem:
    properties:
      created_at:
        example: "2026-04-20T10:00:00Z"
        type: string
      fail_stage:
        example: "null"
        type: string
      is_paid:
        example: true
        type: boolean
      name:
        example: 张三
        type: string
      order_id:
        example: ORD00000000000000000001
        type: string
      order_status:
        example: paid
        type: string
      paid_at:
        example: "2026-04-20T10:00:00Z"
        type: string
      report_id:
        example: 101
        type: integer
      report_type:
        enum:
        - preview
        - full
        example: full
        type: string
      school_name:
        example: 北京大学
        type: string
      status:
        enum:
        - draft
        - unpaid
        - paid
        - collecting
        - planning
        - generating
        - analyzing
        - completed
        - failed
        example: completed
        type: string
      updated_at:
        example: "2026-04-20T11:00:00Z"
        type: string
    type: object
  handler.MPReportProgressDoc:
    properties:
      completion_rate:
        example: 0.4
        type: number
      current_stage:
        example: generating
        type: string
      done_steps:
        example: 4
        type: integer
      total_steps:
        example: 10
        type: integer
    type: object
...
  handler.MPReportStatusData:
    properties:
      fail_stage:
        type: string
      progress:
        $ref: '#/definitions/handler.MPReportProgressDoc'
      report_id:
        example: 101
        type: integer
      report_type:
        example: full
        type: string
      status:
        example: generating
        type: string
    type: object
  handler.MPResponse:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      message:
        example: ok
        type: string
      timestamp:
        example: 1713600000
        type: integer
      user_info:
        $ref: '#/definitions/handler.MPUserInfoDoc'
    type: object
...
  handler.MPUpdateMeData:
    properties:
      user_id:
        example: 42
        type: integer
    type: object
  handler.MPUpdateMeReq:
    properties:
      avatar_url:
        example: https://thirdwx.qlogo.cn/...
        type: string
      nickname:
        example: 张三
        type: string
    type: object
  handler.MPUserInfoDoc:
    properties:
      open_id:
        example: oXXXXXXXXX
        type: string
      user_id:
        example: 42
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: AI 大学生职业规划报告生成系统
  title: XGT Report API
  version: "1.0"
paths:
...
  /mp/auth/bind-phone:
    post:
      consumes:
      - application/json
      description: 首次登录后引导用户授权手机号。小程序调用 wx.getPhoneNumber() 拿到 phone_code，后端解密并 AES-256-GCM
        加密存库，对外返回脱敏手机号。
      parameters:
      - description: wx.login() 临时 code
        in: header
        name: X-Login-Code
        required: true
        type: string
      - description: 系统版本
        in: header
        name: X-System-Version
        required: true
        type: string
      - description: 设备标识
        in: header
        name: X-Device-UUID
        required: true
        type: string
      - description: phone_code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.MPBindPhoneReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.MPResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.MPBindPhoneData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
      summary: 绑定手机号
      tags:
      - mp/auth
  /mp/auth/login:
    post:
      consumes:
      - application/json
      description: 小程序调用 wx.login() 拿到 login_code 放入 Header X-Login-Code，后端 code2session
        换取 openid，查找或新建用户。扫分销商码进入时 body 传 distributor_id。
      parameters:
      - description: wx.login() 返回的临时 code
        in: header
        name: X-Login-Code
        required: true
        type: string
      - description: 系统版本，如 iOS 17.0
        in: header
        name: X-System-Version
        required: true
        type: string
      - description: 设备唯一标识
        in: header
        name: X-Device-UUID
        required: true
        type: string
      - description: 可选：分销商 ID
        in: body
        name: body
        schema:
          $ref: '#/definitions/handler.MPLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.MPResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.MPLoginData'
              type: object
        "400":
          description: 微信 login_code 无效
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
      summary: 微信登录 / 自动注册
      tags:
      - mp/auth
...
  /mp/orders:
    post:
      consumes:
      - application/json
      description: 创建订单并返回微信支付参数。前端用 payment_params 直接调用 wx.requestPayment() 唤起支付。
      parameters:
      - description: wx.login() 临时 code
        in: header
        name: X-Login-Code
        required: true
        type: string
      - description: 系统版本
        in: header
        name: X-System-Version
        required: true
        type: string
      - description: 设备标识
        in: header
        name: X-Device-UUID
        required: true
        type: string
      - description: report_id 和支付金额
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.MPCreateOrderReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.MPResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.MPCreateOrderData'
              type: object
        "400":
          description: amount 不匹配 / 已有待支付订单
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "404":
          description: 报告不存在
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "409":
          description: 重复订单
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
      summary: 创建订单并获取支付参数
      tags:
      - mp/orders
  /mp/orders/detail:
    post:
      consumes:
      - application/json
      description: wx.requestPayment 回调不可靠，支付完成后主动查询确认最终状态。
      parameters:
      - description: wx.login() 临时 code
        in: header
        name: X-Login-Code
        required: true
        type: string
      - description: 系统版本
        in: header
        name: X-System-Version
        required: true
        type: string
      - description: 设备标识
        in: header
        name: X-Device-UUID
        required: true
        type: string
      - description: order_id
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.MPOrderDetailReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.MPResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.MPOrderDetailData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
      summary: 查询订单详情
      tags:
      - mp/orders
...
  /mp/reports:
    post:
      consumes:
      - application/json
      description: 用户填写表单后提交，创建一份新报告，状态为 draft。一个用户可创建多份报告。
      parameters:
      - description: wx.login() 临时 code
        in: header
        name: X-Login-Code
        required: true
        type: string
      - description: 系统版本
        in: header
        name: X-System-Version
        required: true
        type: string
      - description: 设备标识
        in: header
        name: X-Device-UUID
        required: true
        type: string
      - description: 报告表单数据
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.MPCreateReportReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handler.MPResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.MPCreateReportData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
      summary: 创建报告（提交表单）
      tags:
      - mp/reports
  /mp/reports/detail:
    post:
      consumes:
      - application/json
      description: 返回报告所有表单字段 + 状态 + 订单信息，用于回显编辑页面。
      parameters:
      - description: wx.login() 临时 code
        in: header
        name: X-Login-Code
        required: true
        type: string
      - description: 系统版本
        in: header
        name: X-System-Version
        required: true
        type: string
      - description: 设备标识
        in: header
        name: X-Device-UUID
        required: true
        type: string
      - description: report_id
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.MPReportIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MPResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "403":
          description: 无权访问
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "404":
          description: 报告不存在
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
      summary: 获取报告详情（回显表单）
      tags:
      - mp/reports
  /mp/reports/links:
    post:
      consumes:
      - application/json
      description: 一次接口返回所有链接。未付费时 full_h5_url 和 pdf_url 为 null。所有链接为 COS 预签名 URL，有效期
        1 小时。
      parameters:
      - description: wx.login() 临时 code
        in: header
        name: X-Login-Code
        required: true
        type: string
      - description: 系统版本
        in: header
        name: X-System-Version
        required: true
        type: string
      - description: 设备标识
        in: header
        name: X-Device-UUID
        required: true
        type: string
      - description: report_id
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.MPReportIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.MPResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.MPReportLinksData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
      summary: 获取报告链接（预览 H5 / 完整 H5 / PDF）
      tags:
      - mp/reports
  /mp/reports/list:
    post:
      consumes:
      - application/json
      description: 分页返回当前用户的所有报告，按创建时间倒序。每个报告附带最新订单信息。
      parameters:
      - description: wx.login() 临时 code
        in: header
        name: X-Login-Code
        required: true
        type: string
      - description: 系统版本
        in: header
        name: X-System-Version
        required: true
        type: string
      - description: 设备标识
        in: header
        name: X-Device-UUID
        required: true
        type: string
      - description: 分页参数
        in: body
        name: body
        schema:
          $ref: '#/definitions/handler.MPPageReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.MPPageResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/handler.MPPageDataDoc'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/handler.MPReportListItem'
                        type: array
                    type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
      summary: 获取我的报告列表
      tags:
      - mp/reports
  /mp/reports/status:
    post:
      consumes:
      - application/json
      description: 前端每 3~5 秒轮询，直到 status 为 completed 或 failed。progress.completion_rate
        用于进度条。
      parameters:
      - description: wx.login() 临时 code
        in: header
        name: X-Login-Code
        required: true
        type: string
      - description: 系统版本
        in: header
        name: X-System-Version
        required: true
        type: string
      - description: 设备标识
        in: header
        name: X-Device-UUID
        required: true
        type: string
      - description: report_id
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.MPReportIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.MPResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.MPReportStatusData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
      summary: 查询报告生成进度
      tags:
      - mp/reports
...
  /mp/users/me:
    post:
      consumes:
      - application/json
      description: 返回登录用户的全量信息，含报告数、未读消息数、分销商状态。
      parameters:
      - description: wx.login() 临时 code
        in: header
        name: X-Login-Code
        required: true
        type: string
      - description: 系统版本
        in: header
        name: X-System-Version
        required: true
        type: string
      - description: 设备标识
        in: header
        name: X-Device-UUID
        required: true
        type: string
      - description: 空对象
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.MPResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.MPMeData'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
      summary: 获取当前用户信息
      tags:
      - mp/users
  /mp/users/me/update:
    post:
      consumes:
      - application/json
      description: 更新当前登录用户的昵称和头像 URL。
      parameters:
      - description: wx.login() 临时 code
        in: header
        name: X-Login-Code
        required: true
        type: string
      - description: 系统版本
        in: header
        name: X-System-Version
        required: true
        type: string
      - description: 设备标识
        in: header
        name: X-Device-UUID
        required: true
        type: string
      - description: 昵称和头像
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.MPUpdateMeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.MPResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.MPUpdateMeData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.MPErrorResp'
      summary: 更新用户信息（昵称 / 头像）
      tags:
      - mp/users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"